# Fix useless allocs in CommandList