- Add CommandList benchmark
- Fix useless allocs in CommandList