@eNV25 2022-12-02
The runtime performance difference is minimal, but the reduced individual allocations are good.

```
name           old time/op    new time/op    delta
CommandList-4    24.1µs ± 2%    24.6µs ± 3%   +1.96%  (p=0.001 n=9+9)

name           old alloc/op   new alloc/op   delta
CommandList-4      898B ± 4%     1109B ± 1%  +23.48%  (p=0.000 n=10+8)

name           old allocs/op  new allocs/op  delta
CommandList-4      15.0 ± 0%      13.0 ± 0%  -13.33%  (p=0.000 n=10+10)
```